Change answer type to i128 in anticipation of prob 27, constrain other types

+++ rust/src/p0001.rs
@@ -11,8 +11,8 @@ get 3, 5, 6 and 9. The sum of these multiples is 23.
 
 Find the sum of all the multiples of 3 or 5 below 1000.
 */
-pub fn p0001() -> u64 {
-    let mut answer: u64 = 0;
+pub fn p0001() -> i128 {
+    let mut answer: u32 = 0;
     for i in (0..1000).step_by(3) {
         answer += i;
     }
