+++ rust/src/p0010.rs
@@ -12,6 +12,6 @@ Find the sum of all the primes below two million.
 */
 use crate::primes::primes;
 
-pub fn p0010() -> u64 {
+pub fn p0010() -> i128 {
     return primes::<u64>().take_while(|&p| p < 2000000).sum();
 }
